RecommendationAPI/requirements.txt
tornado

pymongo

python-slugify


numpy
pandas==0.17.1

ipython==4.1.1


tornado==4.0.2

flake8==2.5.2

scipy==0.17.0